refactor: playHistory用の型定義(string[][][])がなかったため、新規に型定義
